refactor: PostContent 컴포넌트 수정
